Adds support for transactions
The operation of creating a table in Tableland involves two steps: executing the CREATE statement and inserting an entry into the registry table (system_tables).

Both steps were not inside a transaction.

Added support for transactions to the SQLStore by implementing a Transactor. The Transactor has the knowledge if there is a transaction running and offers the begin-commit-rollback API and also a API for interacting with the db (DBTX).