Adds OpenAPI spec yaml
This is the first version of our API documentation.
It is important to notice that OpenAPI does not support JSON-RPC. To overcome that I added multiple examples to /rpc path.

The best tool for visualizing this spec is probably Swagger UI.
However we are using Gitbook. Gitbook can import parts of the spec. But is very limited.